made a fialing test

+++ Programs/time_proj/time.h
@@ -6,7 +6,7 @@ using namespace std;
 
 struct Time {
     int hour, minute;
-    double second;
+    int second;
 };
